+++ packages/eslint/src/config/react.ts
@@ -1,8 +1,5 @@
 import { isPackageExists } from 'local-pkg';
 
-import { GLOB_SRC } from '../globs';
-import { ensurePackages, interopDefault } from '../utils';
-
 import type {
   OptionsFiles,
   OptionsOverrides,
@@ -11,6 +8,8 @@ import type {
   Rules,
   TypedFlatConfigItem,
 } from '../types';
+import { GLOB_SRC } from '../globs';
+import { ensurePackages, interopDefault } from '../utils';
 
 // react refresh
 const ReactRefreshAllowConstantExportPackages = ['vite'];
@@ -54,7 +53,6 @@ export async function react(
         'react-plugin': plugins['@eslint-react'],
         'react-dom': plugins['@eslint-react/dom'],
         'react-hooks': pluginReactHooks,
-        'react-hooks-extra': plugins['@eslint-react/hooks-extra'],
         'react-naming-convention': plugins['@eslint-react/naming-convention'],
         'react-refresh': pluginReactRefresh,
       },
@@ -113,7 +111,7 @@ export async function react(
         ],
 
         // recommended rules from @eslint-react
-        'react-plugin/no-useless-forward-ref': 'warn',
+        'react-plugin/no-forward-ref': 'warn',
         'react-plugin/no-access-state-in-setstate': 'error',
         'react-plugin/no-array-index-key': 'warn',
         'react-plugin/no-children-count': 'warn',
@@ -130,14 +128,13 @@ export async function react(
         'react-plugin/no-create-ref': 'error',
         'react-plugin/no-direct-mutation-state': 'error',
         'react-plugin/no-duplicate-key': 'error',
-        'react-plugin/no-implicit-key': 'error',
+        'react-plugin/no-implicit-key': 'off',
         'react-plugin/no-missing-key': 'error',
         'react-plugin/no-nested-component-definitions': 'warn',
         'react-plugin/no-redundant-should-component-update': 'error',
         'react-plugin/no-set-state-in-component-did-mount': 'warn',
         'react-plugin/no-set-state-in-component-did-update': 'warn',
         'react-plugin/no-set-state-in-component-will-update': 'warn',
-        'react-plugin/no-string-refs': 'error',
         'react-plugin/no-unsafe-component-will-mount': 'warn',
         'react-plugin/no-unsafe-component-will-receive-props': 'warn',
         'react-plugin/no-unsafe-component-will-update': 'warn',
@@ -187,6 +184,8 @@ export async function react(
         // 禁止循环没有key
         'react/jsx-key': 'warn',
 
+        'react/no-string-refs': 'error',
+
         ...(emotion
           ? {
               'react/no-unknown-property': ['error', { ignore: ['css'] }],
@@ -195,6 +194,7 @@ export async function react(
 
         ...(isTypeAware
           ? {
+              'react-plugin/no-implicit-key': 'error',
               'react-plugin/no-leaked-conditional-rendering': 'warn',
             }
           : {}),